# :enh: bail out early in do_ereport if errstart returns false