Changes `pg_sys::submodules::panic::do_ereport` to bail out early in case `errstart` returns false, in order to lower the performance cost of calling e.g. `debug2!` in a tight loop when neither the log nor the client will report the event.

As postgres versions up to 12 require to pass some palloced data into `errstart`, while postgres 13 and later postpone this to  `errfinish`, in order to maximize performance _and_ legibility, a **copromise** was required, splitting the implementation by version, introducing some duplication.