@eeeebbbbrrrr 2022-12-23
hey, thanks @EdMcBane!  Looks like you need to push up a `cargo fmt` too.  I'll look at this in detail over the weekend.  I was actually thinking about this during my 6h car drive today.