@EdMcBane 2022-12-23
The results are visible but not groundbreaking; tests were conducted with a pair of simple functions:

```rust
#[pg_extern]
fn with_debug() -> i32 {
    debug5!("debug");
    42
}

#[pg_extern]
fn without_debug() -> i32 {
    42
}
```

| Query | Original (ms) | Optimized (ms) |
| ---------| ---------|-----------|
| select i from generate_series(1,10000000) as i where without_debug() < 0;| 1069,830 | N/A |
| select i from generate_series(1,10000000) as i where with_debug() < 0;| 3839,472 |  3030,129 |

Profiling the optimized version shows that most of the remaining time is spent in allocating and deallocating the data that goes into `panic::ErrorReport`, which is done beforehand, and regardless of the eventual return value of `errstart`.

I'm not sure pursuing this further is worthwhile, as the benefits are limited and the impact of changes is significant.