@EdMcBane 2022-12-23
> hey, thanks @EdMcBane! Looks like you need to push up a `cargo fmt` too. 

I was taken aback, then shown a single missing newline-at-end-of-file. Fat fingers :smile: