@eeeebbbbrrrr 2022-12-23
> I'm not sure pursuing this further is worthwhile, as the benefits are limited and the impact of changes is significant.

This is pretty good tho.  And it makes the code correct which is pretty important too.  Thanks a bunch!